Move getAppRoot to utilityFunctions
