+++ src/components/utilityFunctions.ts
@@ -1,3 +1,16 @@
+import path from 'path'
+
+// Determines the folder of the main entry file, as opposed to the
+// project root. Needed for components that traverse the local directory
+// structure (e.g. fileHandler, registerPlugins), as the project
+// root changes its relative path once built.
+export function getAppRootDir() {
+  const serverRoot = String(require.main?.filename)
+  return path.dirname(serverRoot)
+}
+
+// Value equality comparison for arrays and objects, which can
+// be arbitrarily deep
 export const deepEquality = (obj1: any, obj2: any): boolean => {
   if (obj1 === obj2) return true // primitive value check
   const isArray1 = Array.isArray(obj1)
