+++ tsconfig.json
@@ -67,5 +67,5 @@
     "skipLibCheck": true /* Skip type checking of declaration files. */,
     "forceConsistentCasingInFileNames": true /* Disallow inconsistently-cased references to the same file. */
   },
-  "exclude": ["jest.config.js", "./src/plugins/*"]
+  "exclude": ["jest.config.js", "./src/plugins/*", "./database/*"]
 }
